refactor(redux): migrate server and authentication
