+++ src/reducers/dialogs/index.js
@@ -6,18 +6,18 @@
 import { combineReducers } from 'redux';
 
 import appSettingsReducer from '~/features/settings/dialog';
+import authenticationReducer from '~/features/servers/authentication-dialog';
+import deauthenticationReducer from '~/features/servers/deauthentication-dialog';
 import dockDetailsDialogReducer from '~/features/docks/details';
 import featureEditorReducer from '~/features/map-features/editor';
 import geofenceSettingsReducer from '~/features/geofence/slice';
 import savedLocationEditorReducer from '~/features/saved-locations/editor';
+import serverSettingsReducer from '~/features/servers/server-settings-dialog';
 import uavDetailsDialogReducer from '~/features/uavs/details';
 
-import authenticationReducer from './authentication';
-import deauthenticationReducer from './deauthentication';
 import errorHandlingReducer from './error-handling';
 import layerSettingsReducer from './layer-settings';
 import promptReducer from './prompt';
-import serverSettingsReducer from './server-settings';
 
 /**
  * The reducer function that is responsible for handling all dialog-related
