+++ src/views/connections/ConnectionList.jsx
@@ -22,7 +22,7 @@ import React from 'react';
 import { connect } from 'react-redux';
 import TimeAgo from 'react-timeago';
 
-import { showServerSettingsDialog } from '~/actions/server-settings';
+import { showServerSettingsDialog } from '~/features/servers/actions';
 import { listOf } from '~/components/helpers/lists';
 import { ConnectionState } from '~/model/connections';
 import { getConnectionsInOrder } from '~/selectors/ordered';
